+++ ext/CaesarImagesExt.jl
@@ -1,18 +1,38 @@
+module CaesarImagesExt
 
 @info "Loading Caesar tools related to Images.jl."
 
-# using Images
+
+using Images
 # using ImageTransformations
-
 using ColorVectorSpace
-using .Images
-import .GeoPr
-# import GeometricalPredicates as GeoPr
+using UUIDs
+using TensorCast
+using StaticArrays
+using Manifolds
+using DocStringExtensions
+using ProgressMeter
+using Optim
 
-export applyMaskImage
-export makeMaskImage
-export makeMaskImages
-export imhcatPretty
+using Caesar # TODO try reduce to just import Caesar ... below
+import Caesar._PCL as _PCL
+
+# import DistributedFactorGraphs: getManifold
+
+import Base: convert, show
+
+import GeometricalPredicates as GeoPr
+
+import ApproxManifoldProducts: sample, _update!
+import IncrementalInference: getSample, preambleCache, _update!, getManifold
+
+import Caesar: applyMaskImage, makeMaskImage, makeMaskImages, imhcatPretty, toImage
+import Caesar: writevideo, csmAnimationJoinImgs, csmAnimateSideBySide, makeVideoFromData
+import Caesar: overlayScanMatcher
+import Caesar: ScanMatcherPose2, PackedScanMatcherPose2
+import Caesar: _PARCHABLE_PACKED_CLOUD
+import Caesar: ScatterAlign, ScatterAlignPose2, ScatterAlignPose3
+import Caesar: PackedScatterAlignPose2, PackedScatterAlignPose3
 
 """
     $SIGNATURES
@@ -140,6 +160,10 @@ function imhcatPretty(iml::AbstractMatrix{<:Colorant},
   hcat(iml,imll)
 end
 
+include("Images/ImageToVideoUtils.jl")
+include("Images/ScanMatcherUtils.jl")
+include("Images/ScanMatcherPose2.jl")
+include("Images/ScatterAlignPose2.jl")
 
 
-#
+end # module
