+++ ext/Images/ImageToVideoUtils.jl
@@ -1,6 +1,4 @@
 
-export writevideo, csmAnimationJoinImgs, csmAnimateSideBySide
-export makeVideoFromData
 
 
 """
