+++ ext/Images/ScanMatcherPose2.jl
@@ -1,52 +1,6 @@
 
 
-import IncrementalInference: getSample
-import Base: convert, show
-# import DistributedFactorGraphs: getManifold
 
-using .Images
-
-export ScanMatcherPose2, PackedScanMatcherPose2
-
-"""
-$TYPEDEF
-
-This is but one incarnation for how radar alignment factor could work, treat it as a starting point.
-
-Notes
-- Stanard `cvt` argument is lambda function to convert incoming images to user convention of image axes,
-  - default `cvt` flips image rows so that Pose2 xy-axes corresponds to img[x,y] -- i.e. rows down and across from top left corner.
-- Use rescale to resize the incoming images for lower resolution (faster) correlations
-- Both images passed to the construct must have the same type some matrix of type `T`.
-
-Example
--------
-```julia
-arp2 = ScanMatcherPose2(img1, img2, 2) # e.g. 2 meters/pixel 
-```
-
-See also: [`overlayScanMatcher`](@ref)
-"""
-struct ScanMatcherPose2{T} <: IIF.AbstractManifoldMinimize
-  """ reference image for scan matching. """
-  im1::Matrix{T}
-  """ test image to scan match against the reference image. """
-  im2::Matrix{T}
-  """ Common grid scale for both images -- i.e. units/pixel.  
-  Constructor uses two arguments `gridlength`*`rescale=1`=`gridscale`.
-  Arg 0 < `rescale` ≤ 1 is also used to rescale the images to lower resolution for speed. """
-  gridscale::Float64
-
-  # replace inner constructor with transform on image
-  ScanMatcherPose2{T}( im1::AbstractMatrix{T}, 
-                      im2::AbstractMatrix{T},
-                      gridlength::Real,
-                      rescale::Real=1,
-                      cvt = (im)->reverse(Images.imresize(im,trunc.(Int, rescale.*size(im))),dims=1)
-                    ) where {T} = new{T}( cvt(im1), 
-                                          cvt(im2),
-                                          rescale*gridlength )
-end
 
 ScanMatcherPose2(im1::AbstractMatrix{T}, im2::AbstractMatrix{T}, w...) where T = ScanMatcherPose2{T}(im1,im2, w...)
 
@@ -93,7 +47,7 @@ Notes:
 - `tf` is a Manifolds.jl type `::ProductRepr` (or newer `::ArrayPartition`) to represent a `SpecialEuclidean(2)` manifold point.
 """
 function overlayScanMatcher(sm::ScanMatcherPose2, 
-                            trans::AbstractVector{<:Real}=Float64[0;0],
+                            trans::Vector{Float64}=Float64[0;0],
                             rot::Real=0.0;
                             score=Ref(0.0),
                             showscore::Bool=true  )
@@ -119,7 +73,7 @@ end
 #
 
 function overlayScanMatcher(sm::ScanMatcherPose2, 
-                            tf = Manifolds.identity_element(SpecialEuclidean(2));
+                            tf::ArrayPartition = Manifolds.identity_element(SpecialEuclidean(2));
                             kw... )
   #
   M = SpecialOrthogonal(2)
@@ -134,11 +88,6 @@ end
 ## Factor serialization below
 ## =========================================================================================
 
-struct PackedScanMatcherPose2 <: AbstractPackedFactor
-  im1::Vector{Vector{Float64}}
-  im2::Vector{Vector{Float64}}
-  gridscale::Float64
-end
 
 function convert(::Type{<:PackedScanMatcherPose2}, arp2::ScanMatcherPose2)
   TensorCast.@cast pim1[row][col] := arp2.im1[row,col]
